add: renamed action.ts file
